Make timeouts long: they are unsigned ints in C, and only a Java long can properly represent the full range. It would be counter-intuitive to have to set negative timeouts in Java to get longer timeouts in C... Also reorder getVersion() and getStringDescriptorAscii() to their proper places, so they are in the same order as they appear in the libusbx documentation.

+++ src/main/c/src/Transfer.c
@@ -147,11 +147,11 @@ JNIEXPORT jbyte JNICALL METHOD_NAME(Transfer, type)
 }
 
 /**
- * void setTimeout(int)
+ * void setTimeout(long)
  */
 JNIEXPORT void JNICALL METHOD_NAME(Transfer, setTimeout)
 (
-    JNIEnv *env, jobject this, jint timeout
+    JNIEnv *env, jobject this, jlong timeout
 )
 {
     struct libusb_transfer *transfer = unwrapTransfer(env, this);
